# sdk/openapi.json
{
  "components": {
    "schemas": {
...
      "ChainStatus": {
        "description": "Mirrors lib/chain-status.ts's ChainStatus (LEGACY_OFFCHAIN +\nChainStatusBucket) value-for-value \u2014 same reason StatusKey above\nmirrors types.ts: that file already collapses GenLayer's real 14-value\nTransactionStatus into this 4-bucket + legacy shape (via the SDK's own\nisDecidedState(), not a hand-copied list \u2014 see that file's header),\nand duplicating a *different* vocabulary here would let the two\ndrift. scripts/genlayer-read.ts computes the bucket (it has\ngenlayer-js loaded); services/genlayer_indexer.py only ever writes\none of these five strings into a row's chain_status column.\n\nEvery Milestone/Dispute/Prediction row defaults to LEGACY_OFFCHAIN and\nstays there until it's actually linked on-chain (Escrow.contract_address\n/ Prediction.contract_address / Dispute.on_chain_dispute_id set) \u2014 see\nROADMAP.md 4.5's deferred \"add this together with the actual cutover\"\nnote; this is that column, added once there's a real indexer to drive\nit rather than speculatively ahead of one.",
        "enum": [
          "legacy_offchain",
          "processing",
          "decided",
          "finalized",
          "canceled"
        ],
        "title": "ChainStatus",
        "type": "string"
      },
...
      "DisputeCreate": {
        "description": "Body for POST /escrows/{id}/dispute \u2014 escalating a milestone's AI\nverdict to a formal Dispute Court review. `issue` is optional: the\n\"Escalate to Internet Court\" button (components/app/views/\nescrow-detail-view.tsx) fires this with no form of its own, so the\nendpoint fills in a reasonable default from the milestone's most\nrecent ConsensusJob reasoning when omitted \u2014 see routers/escrows.py's\nraise_dispute for that logic. Free-text override kept for any future\ncaller that does want to state its own claim.",
        "properties": {
          "issue": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue"
          }
        },
        "title": "DisputeCreate",
        "type": "object"
      },
...
      "DisputeEnforceRequest": {
        "properties": {
          "approved": {
            "title": "Approved",
            "type": "boolean"
          },
          "ruling": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ruling"
          }
        },
        "required": [
          "approved"
        ],
        "title": "DisputeEnforceRequest",
        "type": "object"
      },
      "DisputeEvidenceCreate": {
        "properties": {
          "description": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "link": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link"
          }
        },
        "required": [
          "description"
        ],
        "title": "DisputeEvidenceCreate",
        "type": "object"
      },
      "DisputeEvidenceRead": {
        "properties": {
          "consensus_job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consensus Job Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "dispute_id": {
            "title": "Dispute Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link"
          },
          "submitter_address": {
            "title": "Submitter Address",
            "type": "string"
          }
        },
        "required": [
          "id",
          "dispute_id",
          "submitter_address",
          "description",
          "created_at"
        ],
        "title": "DisputeEvidenceRead",
        "type": "object"
      },
      "DisputeMessageCreate": {
        "properties": {
          "content": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "DisputeMessageCreate",
        "type": "object"
      },
      "DisputeMessageRead": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dispute_id": {
            "title": "Dispute Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "sender_address": {
            "title": "Sender Address",
            "type": "string"
          }
        },
        "required": [
          "id",
          "dispute_id",
          "sender_address",
          "content",
          "created_at"
        ],
        "title": "DisputeMessageRead",
        "type": "object"
      },
      "DisputeRead": {
        "properties": {
          "adjudication_tx_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adjudication Tx Hash"
          },
          "chain_status": {
            "$ref": "#/components/schemas/ChainStatus",
            "default": "legacy_offchain"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "enforced_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enforced By"
          },
          "escrow_id": {
            "title": "Escrow Id",
            "type": "integer"
          },
          "evidence": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DisputeEvidenceRead"
            },
            "title": "Evidence",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "issue": {
            "title": "Issue",
            "type": "string"
          },
          "messages": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DisputeMessageRead"
            },
            "title": "Messages",
            "type": "array"
          },
          "milestone_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Milestone Id"
          },
          "on_chain_dispute_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "On Chain Dispute Id"
          },
          "on_chain_tx_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "On Chain Tx Hash"
          },
          "opened_by_address": {
            "title": "Opened By Address",
            "type": "string"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "ruling": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ruling"
          },
          "status_key": {
            "$ref": "#/components/schemas/StatusKey"
          }
        },
        "required": [
          "id",
          "escrow_id",
          "opened_by_address",
          "issue",
          "status_key",
          "created_at"
        ],
        "title": "DisputeRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OnChainEvidenceAck": {
        "description": "Body for POST /disputes/{id}/evidence/on-chain \u2014 the frontend\nreporting a tx hash it already got back from signing and sending a\nreal NuanceDisputeCourt.add_evidence transaction (components/app/\ngenlayer-write-client.ts's addEvidenceOnChain). Unlike the off-chain\nsubmit_evidence, this does NOT queue a ConsensusJob \u2014 the real\njudgment happens via adjudicate_dispute on the contract itself\n(triggered automatically by services/genlayer_indexer.py, or\nmanually). `evidence_url` is required (unlike the off-chain path's\noptional `link`) because the contract's own add_evidence only ever\ntakes a URL \u2014 there's nowhere for free-text-only evidence to go\non-chain (see that contract method's own docstring).",
        "properties": {
          "evidence_url": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Evidence Url",
            "type": "string"
          },
          "tx_hash": {
            "title": "Tx Hash",
            "type": "string"
          }
        },
        "required": [
          "tx_hash",
          "evidence_url"
        ],
        "title": "OnChainEvidenceAck",
        "type": "object"
      },
...
      "StatusKey": {
        "enum": [
          "approved",
          "in_review",
          "in_progress",
          "pending",
          "disputed",
          "rejected",
          "cancelled"
        ],
        "title": "StatusKey",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend for Nuance \u2014 AI-validator consensus for claims traditional smart contracts can't settle.",
    "title": "Nuance API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/disputes/{dispute_id}": {
      "get": {
        "operationId": "get_dispute_disputes__dispute_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "title": "Dispute Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dispute",
        "tags": [
          "disputes"
        ]
      }
    },
    "/disputes/{dispute_id}/enforce": {
      "post": {
        "operationId": "enforce_ruling_disputes__dispute_id__enforce_post",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "title": "Dispute Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeEnforceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Enforce Ruling",
        "tags": [
          "disputes"
        ]
      }
    },
    "/disputes/{dispute_id}/evidence": {
      "get": {
        "operationId": "list_evidence_disputes__dispute_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "title": "Dispute Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DisputeEvidenceRead"
                  },
                  "title": "Response List Evidence Disputes  Dispute Id  Evidence Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Evidence",
        "tags": [
          "disputes"
        ]
      },
      "post": {
        "operationId": "submit_evidence_disputes__dispute_id__evidence_post",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "title": "Dispute Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Api-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeEvidenceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeEvidenceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit Evidence",
        "tags": [
          "disputes"
        ]
      }
    },
    "/disputes/{dispute_id}/evidence/on-chain": {
      "post": {
        "description": "The on-chain counterpart to submit_evidence above \u2014 reached once\ncomponents/app/genlayer-write-client.ts's addEvidenceOnChain has\nalready signed and sent a real NuanceDisputeCourt.add_evidence\ntransaction. Added 2026-09-08 to close the actual gap that caused a\nreal bug: before this endpoint existed, submitting evidence for ANY\ndispute \u2014 on-chain or not \u2014 always went through submit_evidence\nabove, which always queues an off-chain ConsensusJob. That silently\nrouted an on-chain-filed dispute's ruling through Nuance's own\noff-chain AI review instead of real GenVM validators, with nothing\nin the UI making the swap visible (found live, see contracts/\nnuance_dispute_court.py's add_evidence docstring for the full\naccount).\n\nUnlike submit_evidence, this does NOT queue a ConsensusJob \u2014 real\njudgment happens via adjudicate_dispute on the contract itself\n(services/genlayer_indexer.py's trigger_pending_adjudications\ntriggers it automatically once the dispute is open and matched).\nStill creates a local DisputeEvidence row purely so the evidence\nshows up in the UI's evidence list \u2014 same \"not a trust boundary\"\nreasoning as every other on-chain ack in this app: nothing here\nverifies the hash is real or that the contract call actually\nsucceeded; only the indexer reading the contract's own ruling back\nis what actually matters.",
        "operationId": "submit_evidence_on_chain_disputes__dispute_id__evidence_on_chain_post",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "title": "Dispute Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnChainEvidenceAck"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeEvidenceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit Evidence On Chain",
        "tags": [
          "disputes"
        ]
      }
    },
    "/disputes/{dispute_id}/messages": {
      "get": {
        "operationId": "list_messages_disputes__dispute_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "title": "Dispute Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DisputeMessageRead"
                  },
                  "title": "Response List Messages Disputes  Dispute Id  Messages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Messages",
        "tags": [
          "disputes"
        ]
      },
      "post": {
        "operationId": "send_message_disputes__dispute_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "title": "Dispute Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeMessageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeMessageRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send Message",
        "tags": [
          "disputes"
        ]
      }
    },
...
    "/escrows/{escrow_id}/dispute": {
      "post": {
        "description": "Escalates the escrow's active milestone to a formal Dispute Court\nreview. Not previously implemented anywhere in this backend \u2014 ROADMAP.md\n3.1's Part 1 checklist marks \"GET/POST /disputes (implicit via escrow)\"\ndone, but no such endpoint, implicit or otherwise, actually existed;\nescrow-detail-view.tsx's \"Escalate to Internet Court\" button has been\nunwired since it was written. This is that missing endpoint, escrow-\nscoped to match the \"implicit via escrow\" framing rather than a bare\nPOST /disputes.\n\nEither party to the escrow may open one \u2014 mirrors contracts/\nnuance_dispute_court.py's file_dispute (claimant is whoever calls it;\nthe other party is derivable from the escrow, same as this repo's\nexisting mapDispute on the frontend already does \u2014 no separate\n`respondent` field needed). Creates the Dispute row and an initial\nDisputeEvidence entry together, then queues the same AI-jury review\nsubmit_evidence already runs below \u2014 a claim with no supporting text\nisn't reviewable.",
        "operationId": "raise_dispute_escrows__escrow_id__dispute_post",
        "parameters": [
          {
            "in": "path",
            "name": "escrow_id",
            "required": true,
            "schema": {
              "title": "Escrow Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeCreateRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Raise Dispute",
        "tags": [
          "escrows"
        ]
      }
    },
...
  }
}